testweb/requirements.txt
django
drf-yasg
django-rest-swagger
djangorestframework
pyyaml
django-filter
